DOC: Archive project
This repository is obsolete as of January 2025.

Previously, this repository was deployed on Heroku[^1] and configured as a
webhook to trigger CircleCI builds for the `Slicer/apidocs.slicer.org` GitHub
project whenever updates were made to the Slicer repository.

It has been superseded by the `trigger-doxygen-build-and-publish.yml` workflow,
introduced in Slicer pull request [#8144](https://github.com/Slicer/Slicer/pull/8144)
and maintained within the `Slicer` repository. This new workflow triggers the
`doxygen-build-and-publish.yml` workflow, which is maintained in the `apidocs.slicer.org` repository.